fix(upload): 修正分片上传中附件ID的数据类型

+++ datasophon-api/src/main/java/com/datasophon/api/controller/TempFileController.java
@@ -71,7 +71,7 @@ public class TempFileController extends ApiController {
     public Result uploadChunk(
             @NotNull(message = "分片不能为空") @RequestPart("chunk") MultipartFile chunk,
             @NotNull(message = "chunkNo不能为空") @Schema(description = "分片索引，0-base") Integer chunkNo,
-            @NotNull(message = "attachId不能为空") Long attachId,
+            @NotNull(message = "attachId不能为空") Integer attachId,
             String md5) {
         ChunkDTO info = new ChunkDTO();
         info.setChunk(chunk);
